backend/requirements.txt
Flask>=2.0.1
Werkzeug>=2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz
flask-cors>=3.0.10